Rename PlayfabGSDK to PlayFabGSDK
