+++ UnrealPlugin/Source/PlayfabGSDK/Private/GSDKConfiguration.cpp
@@ -15,7 +15,7 @@
 #include "GSDKConfiguration.h"
 
 #include "GSDKInternalUtils.h"
-#include "PlayfabGSDK.h"
+#include "PlayFabGSDK.h"
 #include "Logging/LogMacros.h"
 #include "Misc/FileHelper.h"
