+++ UnrealPlugin/Source/PlayfabGSDK/Private/GSDKUtils.cpp
@@ -14,7 +14,7 @@
 
 #include "GSDKUtils.h"
 
-#include "PlayfabGSDK.h"
+#include "PlayFabGSDK.h"
 
 const FGameServerConnectionInfo UGSDKUtils::GetGameServerConnectionInfo()
 {
