+++ UnrealPlugin/Source/PlayfabGSDK/Private/PlayFabGSDK.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "PlayfabGSDK.h"
+#include "PlayFabGSDK.h"
 
 #if UE_SERVER
 	#define PLAYFAB_GSDK_SERVER true
