+++ UnrealPlugin/Source/PlayfabGSDK/Public/PlayFabGSDK.h
